MyWebServer
This program is able to send the files to the client's browser with an
appropriate header file which declares the File type and length to the
bowser. The client will connect to this MyWebServer program using the
browser and navigating to localhost:2540. The program will display the
contents of the folder in which this program is located and if the
client tries to navigate to the parent folder the program won't allow
the client to do so or security reasons. The program is capable of
handling a HTML form but I still need to work on it.