Mimer
In this assignment the main objective was to use the MyWebServer program
and create a new handler which will handle a user defined file type(.xyz
in this case). This assignment taught me how to create and register a
new file extension type in windows and also how to associate it with a
batch file that handles these types of file extensions. This assignment
also included Marshalling of data in XML format using a back channel.
The client will connect to the MyWebServer program and will select a
file of .xyz extension which will call the associated batch file that
handles these kinds of extension and display its content. The
modification that I did will call the back channel program that will
marshall the data to the MyWebServer program in the XML format and will
also create a temporary output file called mimer.output in the temp
folder where the data which has been marshalled will be located.